// lab7/frontend/src/api/Api.ts
export interface SerializerUserJSON {
  id?: string;
  is_moderator?: boolean;
  login?: string;
  password?: string;
}